requirements.txt
alabaster
annotated-types
anyio
appdirs
appnope
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-lru
attrs
babel
beautifulsoup4
bleach
boto3
botocore
certifi
cffi
charset-normalizer
colorama
comm
contourpy
cssselect
cycler
debugpy
decorator
defusedxml
distro
docutils
et-xmlfile
exceptiongroup
executing
fastjsonschema
filelock
fonttools
fqdn
frozendict
fsspec
gitdb
gitpython
gpt4all
h11
html5lib
httpcore
httpx
huggingface-hub
idna
imagesize
iniconfig
ipykernel
ipython
ipywidgets
isoduration
jedi
jinja2
jmespath
json5
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
kiwisolver
lxml
markupSafe
matplotlib
matplotlib-inline
mistune
mpmath
multitasking
nbclient
nbconvert
nbformat
nest-asyncio
networkx
notebook
notebook_shim
numpy
numpydoc
openai
openpyxl
overrides
packaging
pandas
pandocfilters
parse
parso
peewee
pexpect
pillow
platformdirs
pluggy
prometheus_client
prompt_toolkit
psutil
ptyprocess
pure-eval
py-cpuinfo
pycparser
pydantic
pydantic_core
pydispatcher
pydocstyle
pygments
pyparsing
pyquery
pytest
python-dateutil
python-json-logger
pytz
pyyaml
pyzmq
qtconsole
qtpy
referencing
regex
requests
rfc3339-validator
rfc3986-validator
rpds-py
s3transfer
safetensors
scipy
send2trash
sentencepiece
six
smmap
sniffio
snowballstemmer
soupsieve
sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
stack-data
sympy
tabulate
terminado
tifffile
tinycss2
tokenizers
tomli
torch
torchvision
tornado
tqdm
traitlets
transformers
types-python-dateutil
typing_extensions
tzdata
uri-template
urllib3
wcwidth
webcolors
webencodings
websocket-client
widgetsnbextension
yfinance
scikit-learn